requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
psycopg2-binary==2.9.7
pylint==2.17.5
black==23.7.0
django-cors-headers==4.2.0
django-filter==23.2
django-environ==0.10.0
django-pgconnection==1.0.3
django-extensions==3.2.3
django-cors-headers==4.2.0
pre-commit==3.3.3
pytest-django==4.5.2
drf-yasg==1.21.7
moto==4.2.0
boto3==1.28.36
django-filter==23.2
django-rq==2.8.1
diskcache==5.6.3
redis==4.6.0
jsonrpcclient==4.0.3
backoff==2.2.1
python-dateutil
